refactor: factor out common arguments
This allows to avoid duplicating clap config and comments
